# Measure emitting records counts and their sizes for input and output metrics and used it in in_monitor_agent (revised)