--- a/lib/fluent/plugin/input.rb
+++ b/lib/fluent/plugin/input.rb
@@ -34,6 +34,13 @@ module Fluent
         @emit_records = 0
         @emit_size = 0
         @counter_mutex = Mutex.new
+        @enable_size_metrics = false
+      end
+
+      def configure(conf)
+        super
+
+        @enable_size_metrics = !!system_config.enable_size_metrics
       end
 
       def statistics
@@ -48,7 +55,7 @@ module Fluent
       def metric_callback(es)
         @counter_mutex.synchronize do
           @emit_records += es.size
-          @emit_size += es.to_msgpack_stream.bytesize
+          @emit_size += es.to_msgpack_stream.bytesize if @enable_size_metrics
         end
       end