--- a/lib/fluent/plugin/output.rb
+++ b/lib/fluent/plugin/output.rb
@@ -801,8 +801,10 @@ module Fluent
         @counter_mutex.synchronize{ @emit_count += 1 }
         begin
           process(tag, es)
-          @counter_mutex.synchronize{ @emit_records += es.size }
-          @counter_mutex.synchronize{ @emit_size += es.to_msgpack_stream.bytesize }
+          @counter_mutex.synchronize do
+            @emit_records += es.size
+            @emit_size += es.to_msgpack_stream.bytesize
+          end
         rescue
           @counter_mutex.synchronize{ @num_errors += 1 }
           raise
@@ -968,8 +970,10 @@ module Fluent
         write_guard do
           @buffer.write(meta_and_data, enqueue: enqueue)
         end
-        @counter_mutex.synchronize{ @emit_records += records }
-        @counter_mutex.synchronize{ @emit_size += es.to_msgpack_stream.bytesize }
+        @counter_mutex.synchronize do
+          @emit_records += records
+          @emit_size += es.to_msgpack_stream.bytesize
+        end
         true
       end
 
@@ -986,8 +990,10 @@ module Fluent
         write_guard do
           @buffer.write(meta_and_data, format: format_proc, enqueue: enqueue)
         end
-        @counter_mutex.synchronize{ @emit_records += records }
-        @counter_mutex.synchronize{ @emit_size += es.to_msgpack_stream.bytesize }
+        @counter_mutex.synchronize do
+          @emit_records += records
+          @emit_size += es.to_msgpack_stream.bytesize
+        end
         true
       end
 
@@ -1012,8 +1018,10 @@ module Fluent
         write_guard do
           @buffer.write({meta => data}, format: format_proc, enqueue: enqueue)
         end
-        @counter_mutex.synchronize{ @emit_records += records }
-        @counter_mutex.synchronize{ @emit_size += es.to_msgpack_stream.bytesize }
+        @counter_mutex.synchronize do
+          @emit_records += records
+          @emit_size += es.to_msgpack_stream.bytesize
+        end
         true
       end