--- a/lib/fluent/root_agent.rb
+++ b/lib/fluent/root_agent.rb
@@ -315,6 +315,7 @@ module Fluent
       # See also 'fluentd/plugin/input.rb'
       input.context_router = @event_router
       input.configure(conf)
+      @event_router.add_metric_callbacks(input.plugin_id, Proc.new {|es| input.metric_callback(es) })
       @inputs << input
 
       input