--- a/lib/fluent/supervisor.rb
+++ b/lib/fluent/supervisor.rb
@@ -570,6 +570,7 @@ module Fluent
         suppress_repeated_stacktrace: true,
         ignore_repeated_log_interval: nil,
         without_source: nil,
+        enable_input_metrics: nil,
         use_v1_config: true,
         strict_config_value: nil,
         supervise: true,