--- a/lib/fluent/supervisor.rb
+++ b/lib/fluent/supervisor.rb
@@ -571,6 +571,7 @@ module Fluent
         ignore_repeated_log_interval: nil,
         without_source: nil,
         enable_input_metrics: nil,
+        enable_size_metrics: nil,
         use_v1_config: true,
         strict_config_value: nil,
         supervise: true,