--- a/lib/fluent/system_config.rb
+++ b/lib/fluent/system_config.rb
@@ -28,7 +28,7 @@ module Fluent
       :without_source, :rpc_endpoint, :enable_get_dump, :process_name,
       :file_permission, :dir_permission, :counter_server, :counter_client,
       :strict_config_value, :enable_msgpack_time_support, :disable_shared_socket,
-      :metrics
+      :metrics, :without_input_metrics
     ]
 
     config_param :workers,   :integer, default: 1
@@ -47,6 +47,7 @@ module Fluent
     config_param :strict_config_value, :bool, default: nil
     config_param :enable_msgpack_time_support, :bool, default: nil
     config_param :disable_shared_socket, :bool, default: nil
+    config_param :without_input_metrics, :bool, default: true
     config_param :file_permission, default: nil do |v|
       v.to_i(8)
     end