… test/test_root_agent.rb