<!--
Thank you for contributing to Fluentd!
Your commits need to follow DCO: https://probot.github.io/apps/dco/
And please provide the following information to help us make the most of your pull request:
-->

**Which issue(s) this PR fixes**: 
None

**What this PR does / why we need it**: 
Currently, Fluentd metrics do not have emit size for calculating flow rate (per sec or  per hour).
Fluentd Input plugin and Output should have `emit_records` and `emit_size` metrics.

But, currently, I disabled input metrics and size metrics by default for performance concerns.

`enable_input_metrics` in system config is `false` by default. 
`enable_size_metrics` in system config is also `false` by default.


**Docs Changes**:
Needed.

**Release Note**: 

Same as title.